Fixed a couple of bugs in triplifier: Hey, John,
Just reporting a couple of bugs in the command-line triplifer that I suspect might be easily fixed.

1.  If I run the command-line triplifier without specifying an output directory, I get

Exception in thread "main" java.io.FileNotFoundException: Unable to create directory /path/to/tripleOutput
    at settings.PathManager.setPaths(PathManager.java:60)
    at settings.PathManager.setDirectory(PathManager.java:17)
    at triplify.main(triplify.java:114)

Yet it actually does create the directory, and if I then run the same command again, it works.  Weird.

2.  The help output says that "dwc" is a simplifier option, but if I try to use it, I get "dwc not a valid simplifier type".